requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosmtplib==4.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.32.0
azure-storage-blob==12.25.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.37.15
botocore==1.37.15
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
customtkinter==5.2.2
darkdetect==0.8.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
Flask==3.1.0
Flask-Sockets==0.2.1
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-speech==2.31.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.21
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langsmith==0.3.16
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
msal==1.32.0
multidict==6.2.0
narwhals==1.31.0
nodemon==0.0.17
numpy==2.2.4
oauthlib==3.2.2
openai==1.66.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pygit2==1.17.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.3
PyNaCl==1.5.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.12.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
razorpay==1.4.2
recordclass==0.22.1
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
setuptools==76.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
social-auth-core==4.5.6
sounddevice==0.5.1
SQLAlchemy==2.0.39
starlette==0.46.1
streamlit==1.43.2
stripe==11.6.0
telnyx==2.1.5
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
